# res/openapi.yaml
openapi: 3.0.0
info:
  title: grommunio Admin API
  version: 1.0.0
servers:
  - url: '/api/v1'
    description: Default deployment path
paths:
...
  /system/dashboard:
    get:
      summary: Get System dashboard data
      tags:
        - System Admin/Dashboard
      security:
        - JWTCookie: []
      responses:
        '200':
          description: Data returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  disks:
                    description: List of disks
                    type: array
                    items:
                      type: object
                      description: Disk statistics.
                      properties:
                        percent:
                          type: number
                          description: Percentage of used memory
                        total:
                          description: Total disk space (bytes)
                          type: integer
                        used:
                          description: Used disk space (bytes)
                          type: integer
                        free:
                          description: Free disk space (bytes)
                          type: integer
                        device:
                          type: string
                        mountpoint:
                          type: string
                        filesystem:
                          type: string
                  load:
                    type: array
                    description: Load averages (1 / 5 / 15 minutes)
                    items:
                      type: number
                  cpuPercent:
                    description: Current CPU usage in percent
                    type: object
                    properties:
                      user:
                        type: number
                        description: Percentage of user CPU time
                      system:
                        type: number
                        description: Percentage of system CPU time
                      io:
                        type: number
                        description: Percentage of CPU time used for IO operations
                      interrupt:
                        type: number
                        description: Percentage of CPU time used by interrupts
                      steal:
                        type: number
                        description: Percentage of CPU time spent waiting for the hypervisor
                      idle:
                        type: number
                        description: Idle CPU time
                  memory:
                    description: RAM statistics
                    type: object
                    properties:
                      percent:
                        description: Percent of used memory
                        type: number
                      total:
                        description: Total memory (bytes)
                        type: integer
                      free:
                        description: Unused memory (bytes)
                        type: integer
                      used:
                        description: Memory used by applications (bytes)
                        type: integer
                      buffer:
                        description: Memory used for buffers (bytes)
                        type: integer
                      cache:
                        description: Memory used for cached data (bytes)
                        type: integer
                      available:
                        description: Memory that is available for applications (bytes)
                  swap:
                    description: Swap statistics
                    type: object
                    properties:
                      percent:
                        description: Percentage of used swap memory
                        type: number
                      total:
                        description: Total swap memory (bytes)
                        type: integer
                      free:
                        description: Available swap memory (bytes)
                        type: integer
                      used:
                        description: Used swap memory (bytes)
                        type: integer
                  booted:
                    allOf:
                      - $ref: '#/components/schemas/dateTime'
                      - description: Time the machine was booted
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/dashboard/services:
    get:
      summary: Get list of services
      tags:
        - System Admin/Dashboard
      security:
        - JWTCookie: []
      responses:
        '200':
          description: List of services returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    description: List of systemd services
                    type: array
                    items:
                      $ref: '#/components/schemas/service'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/dashboard/services/{unit}:
    get:
      summary: Get information about a specific service
      tags:
        - System Admin/Dashboard
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/systemdUnit'
      responses:
        '200':
          description: Inforamtion about service returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/service'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/dashboard/services/{unit}/{action}:
    post:
      summary: Send signal to a service
      tags:
        - System Admin/Dashboard
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/systemdUnit'
        - name: action
          description: Signal to send
          in: path
          required: true
          schema:
            type: string
            enum: [start, stop, restart, reload, enable, disable]
      responses:
        '201':
          description: Signal was sent
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /system/license:
    get:
      summary: Get information about the currently installed License
      tags:
        - System Admin/License
      security:
        - JWTCookie: []
      responses:
        '200':
          description: License information returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/license'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    put:
      summary: Upload new license file
      tags:
        - System Admin/License
      security:
        - JWTCookie: []
      responses:
        '200':
          description: License updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/license'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/license/certificate.pem:
    get:
      description: Download the certificate file
      tags:
        - System Admin/License
      security:
        - JWTCookie: []
      responses:
        '200':
          description: License data returned
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /system/antispam/{path}:
    get:
      summary: Make call to grommunio-antispam
      description: Proxies any request including query parameters to the grommunio-antispam backend
      tags:
        - System Admin/Proxies
      security:
        - JWTCookie: []
      parameters:
        - name: path
          in: path
          description: grommunio-antispam endpoint to query
          required: true
          schema:
            type: string
      responses:
        '403':
          description: Insufficient privileges or endpoint not allowed
        '503':
          description: Connection to grommunio-antispam failed
        default:
          description: Response returned by the grommunio-antispam backend
  /system/vhostStatus:
    get:
      summary: Get list of vhosts
      tags:
        - System Admin/Proxies
      security:
        - JWTCookie: []
      responses:
        '200':
          description: List returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/vhostStatus/{host}:
    get:
      summary: Make call nginx vhost status endpoint
      tags:
        - System Admin/Proxies
      security:
        - JWTCookie: []
      parameters:
        - name: host
          in: path
          description: Host to query
          required: true
          schema:
            type: string
      responses:
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Connection to vhost failed
        default:
          description: Response returned by the vhost status endpoint
  /system/mailq:
    get:
      summary: Retrieve mailq output
      tags:
        - Misc
      security:
        - JWTCookie: []
      responses:
        '200':
          description: Output returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  postfixMailq:
                    description: Output of the postfix mailq command
                    type: string
                  gromoxMailq:
                    description: Output of the gromox-mailq command
                    type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /system/cli:
    post:
      summary: Remote CLI invocation
      tags:
        - Misc
      security:
        - JWTCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command:
                  type: string
                  description: CLI command to execute
                mode:
                  type: string
                  description: CLI mode (execute or complete)
                  enum: [exec, complete]
                  default: exec
                color:
                  type: boolean
                  description: Enable terminal colors
                  default: false
                fs:
                  type: object
                  nullable: true
                  description: If not null, enable filesystem redirection. Each property is a file path
                  additionalProperties:
                    type: string
                    description: Content of the file
      responses:
        '200':
          description: Execution successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: Command exit code (`exec` mode only)
                  stdout:
                    type: string
                    description: Command output (`exec` mode only)
                  fs:
                    type: object
                    nullable: true
                    additionalProperties:
                      type: object
                      properties:
                        mode:
                          type: string
                          description: Mode the file was opened with
                        conf:
                          type: string
                          nullable: true
                          description: Configuration key the file path was derived from
                        content:
                          type: string
                          description: Content of the file
                  completions:
                    type: array
                    description: Possible command completions (`complete` mode only)
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/sync/top:
    get:
      summary: Get grommunio-sync usage information
      description: Get current usage. Should be called at least twice to provide useful information.
      tags:
        -  Misc
      security:
        - JWTCookie: []
      parameters:
        - name: filterUpdated
          in: query
          description: Maximum number of seconds since the last update
          schema:
            type: integer
        - name: filterEnded
          in: query
          description: Maximum number of seconds since the process ended
          schema:
            type: integer
      responses:
        '200':
          description: Data returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  maxUpdated:
                    type: integer
                    description: Maximum effective value for `filterUpdated`
                  maxEnded:
                    type: integer
                    description: Maximum effective value for `filterEnded`
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        pid:
                          type: integer
                          description: Process ID
                        ip:
                          type: string
                          description: IP address of the device
                        user:
                          type: string
                        start:
                          type: integer
                          description: UNIX timestamp of process start
                        devtype:
                          type: string
                        devid:
                          type: string
                        devagent:
                          type: string
                        command:
                          type: integer
                        ended:
                          type: integer
                          description: UNIX timestamp of process end or 0 if still running
                        push:
                          type: boolean
                        addinfo:
                          type: string
                          description: Additional information about the process
                        update:
                          type: integer
                          description: UNIX timestamp of last update
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    JWTCookie:
      type: apiKey
      in: cookie
      name: grommunioAuthJwt
  parameters:
...
    systemdUnit:
      name: unit
      description: Name of the unit to signal
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    dateTime:
      type: string
      description: Date string with time
      pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
      nullable: True
...
    service:
      type: object
      properties:
        state:
          type: string
        substate:
          type: string
        description:
          type: string
          nullable: true
        autostart:
          type: string
          description: State of the unit file
        since:
          $ref: '#/components/schemas/dateTime'
        name:
          type: string
          description: Name of the service
        unit:
          type: string
          description: Name of the systemd unit. Required for action endpoints.
...
    license:
      type: object
      properties:
        product:
          type: string
          nullable: true
          description: Product name
        maxUsers:
          type: integer
          description: License user limit
        notBefore:
          $ref: '#/components/schemas/dateTime'
        notAfter:
          $ref: '#/components/schemas/dateTime'
        currentUsers:
          type: integer
          nullable: true
          description: Number of currently existing users
        certificate:
          type: string
          nullable: true
          description: Download link for the certificate
...
  responses:
    ServerError:
      description: An error occured while processing the request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: String representation of the exception
    InvalidRequest:
      description: Validation of input parameters failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                description: List of errors encountered during validation
                items:
                  type: string
    DatabaseError:
      description: The database query failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Description of the error
    ServiceUnavailable:
      description: One of the required external services is unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Description of the error
    NotFound:
      description: The requested resource could not be found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...